requirements.txt
python-dotenv
requests
openai
notion-client
gspread
oauth2client
